# functions/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Organizr Custom GPT Actions",
    "description": "GPT-safe endpoints for checking user readiness and creating extractors from Gmail messages.",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "https://api-45fybwj2xq-uc.a.run.app"
    }
  ],
  "security": [
    {
      "CustomGptOAuth": [
        "profile",
        "gmail",
        "extractors"
      ]
    }
  ],
  "paths": {
...
    "/api/gpt/extractor-from-analysis": {
      "post": {
        "description": "Persist an extractor using schema and parser analysis generated by Custom GPT.",
        "operationId": "createExtractorFromCustomGptAnalysis",
        "deprecated": false,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "subject": {
                    "type": "string"
                  },
                  "extractorId": {
                    "type": "string",
                    "description": "Optional existing extractor id. When provided, the analyzed subject is attached to that extractor instead of creating a new one."
                  },
                  "emails": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "validationSample": {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "body": {
                        "type": "string"
                      },
                      "subject": {
                        "type": "string"
                      },
                      "from": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "body"
                    ]
                  },
                  "analysis": {
                    "type": "object",
                    "additionalProperties": true,
                    "properties": {
                      "explanation": {
                        "type": "string"
                      },
                      "schemaFields": {
                        "type": "array",
                        "items": {
                          "type": "object"
                        }
                      },
                      "scriptCode": {
                        "type": "string"
                      },
                      "validationSample": {
                        "type": "object",
                        "additionalProperties": false,
                        "properties": {
                          "body": {
                            "type": "string"
                          },
                          "subject": {
                            "type": "string"
                          },
                          "from": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "body"
                        ]
                      },
                      "sampleExtractedResults": {
                        "type": "array",
                        "items": {
                          "type": "object"
                        }
                      }
                    },
                    "required": [
                      "schemaFields",
                      "scriptCode"
                    ]
                  }
                },
                "required": [
                  "subject",
                  "analysis"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Extractor created or subject attached to an existing extractor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateExtractorFromSubjectResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Custom GPT analysis.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GptActionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired OAuth bearer token."
          },
          "404": {
            "description": "The requested extractor was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GptActionResponse"
                }
              }
            }
          },
          "409": {
            "description": "The subject already exists on the target extractor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GptActionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Custom GPT persistence orchestration failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GptActionResponse"
                }
              }
            }
          },
          "502": {
            "description": "Extractor persistence failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GptActionResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "CustomGptOAuth": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://api-45fybwj2xq-uc.a.run.app/oauth/authorize",
            "tokenUrl": "https://api-45fybwj2xq-uc.a.run.app/oauth/token",
            "scopes": {
              "profile": "Identify the signed-in Organizr user.",
              "gmail": "Use the user's connected Gmail data for requested searches.",
              "extractors": "Create and update extractors owned by the user."
            }
          }
        }
      }
    },
    "schemas": {
      "GptActionCode": {
        "type": "string",
        "enum": [
          "READY",
          "AUTH_REQUIRED",
          "GMAIL_CONNECTION_REQUIRED",
          "LLM_PROVIDER_REQUIRED",
          "TICKET_DESCRIPTION_REQUIRED",
          "SUBJECT_REQUIRED",
          "NO_EMAILS_FOUND",
          "CUSTOM_GPT_ANALYSIS_REQUIRED",
          "MANUAL_PAYLOAD_REQUIRED",
          "PENDING_MANUAL_COMPLETION",
          "ANALYSIS_FAILED",
          "EXTRACTOR_NOT_FOUND",
          "EXTRACTOR_CREATE_FAILED",
          "EXTRACTOR_READ_FAILED",
          "EXTRACTOR_EDIT_FAILED",
          "EXTRACTOR_UPDATE_FAILED",
          "SUBJECT_ALREADY_EXISTS",
          "UNEXPECTED_ERROR"
        ]
      },
      "LlmProviderCapability": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "available": {
            "type": "boolean"
          },
          "source": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "user",
              "server",
              "chatgpt",
              null
            ]
          },
          "actionCode": {
            "type": [
              "string",
              "null"
            ]
          },
          "actionUrl": {
            "type": "string"
          }
        },
        "required": [
          "available",
          "source",
          "actionCode",
          "actionUrl"
        ]
      },
      "GptCapabilities": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "authenticated": {
            "type": "boolean"
          },
          "gmailConnected": {
            "type": "boolean"
          },
          "llmAvailable": {
            "type": "boolean"
          },
          "defaultProvider": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "auto",
              "gemini",
              "openai",
              "customgpt",
              null
            ]
          },
          "providers": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "gemini": {
                "$ref": "#/components/schemas/LlmProviderCapability"
              },
              "openai": {
                "$ref": "#/components/schemas/LlmProviderCapability"
              },
              "customgpt": {
                "$ref": "#/components/schemas/LlmProviderCapability"
              }
            }
          },
          "availableProviders": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "gemini",
                "openai",
                "customgpt"
              ]
            }
          },
          "supportedProviderModes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "auto",
                "gemini",
                "openai",
                "customgpt"
              ]
            }
          },
          "actionUrl": {
            "type": "string",
            "format": "uri"
          },
          "nextSteps": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "authenticated",
          "gmailConnected",
          "llmAvailable",
          "defaultProvider",
          "providers",
          "availableProviders",
          "supportedProviderModes",
          "actionUrl",
          "nextSteps"
        ]
      },
      "GptActionResponse": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "code": {
            "$ref": "#/components/schemas/GptActionCode"
          },
          "message": {
            "type": "string"
          },
          "capabilities": {
            "$ref": "#/components/schemas/GptCapabilities"
          }
        },
        "required": [
          "ok",
          "code",
          "message"
        ]
      },
...
      "CreateExtractorFromSubjectResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/GptActionResponse"
          },
          {
            "type": "object",
            "properties": {
              "subject": {
                "type": "string"
              },
              "mode": {
                "type": "string",
                "enum": [
                  "create-extractor",
                  "attach-subject",
                  "customgpt-manual-analysis",
                  "llm-analysis"
                ]
              },
              "emailCount": {
                "type": "integer"
              },
              "emails": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "targetExtractor": {
                "type": "string"
              },
              "expectedAnalysisShape": {
                "type": "object",
                "additionalProperties": true
              },
              "created": {
                "type": "boolean"
              },
              "targetExtractorUpdated": {
                "type": "boolean"
              },
              "schemaChanged": {
                "type": "boolean"
              },
              "extractor": {
                "type": "object",
                "description": "Persisted extractor owned by the authenticated user."
              }
            },
            "required": [
              "subject"
            ]
          }
        ]
      },
...
    }
  }
}